requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
APScheduler==3.10.4
python-dotenv==1.0.0